.history/requirements_20251101215446.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1